refactor: separate login controller, service for clearer auth flow
