+++ src/main/java/com/example/devSns/Login/LoginService.java
@@ -1,0 +1,32 @@
+package com.example.devSns.Login;
+
+import com.example.devSns.Jwt.JwtUtil;
+import com.example.devSns.Member.Member;
+import com.example.devSns.Member.MemberRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional
+@RequiredArgsConstructor
+public class LoginService {
+
+    private final MemberRepository memberRepository;
+    private final PasswordEncoder passwordEncoder;
+    private final JwtUtil jwtUtil;
+
+    public String login(String email, String rawPassword) {
+
+        Member member = memberRepository.findByEmail(email)
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 이메일입니다."));
+
+        if (!passwordEncoder.matches(rawPassword, member.getPassword())) {
+            throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
+        }
+
+        // JWT 발급
+        return jwtUtil.createToken(member.getId());
+    }
+}
